refactor wave detail router
