+++ src/app/router_registry.py
@@ -204,6 +204,7 @@ from app.routers.dpm_wave_campaign_workflow import (
 from app.routers.dpm_wave_campaign_workflow_boards import (
     router as dpm_wave_campaign_workflow_boards_router,
 )
+from app.routers.dpm_wave_detail import router as dpm_wave_detail_router
 from app.routers.dpm_wave_evidence import router as dpm_wave_evidence_router
 from app.routers.dpm_wave_items import router as dpm_wave_items_router
 from app.routers.dpm_wave_lifecycle_actions import router as dpm_wave_lifecycle_actions_router
@@ -446,6 +447,7 @@ def register_routers(app: FastAPI) -> None:
     app.include_router(dpm_proof_pack_ai_router)
     app.include_router(dpm_waves_router)
     app.include_router(dpm_wave_lookup_router)
+    app.include_router(dpm_wave_detail_router)
     app.include_router(dpm_wave_items_router)
     app.include_router(dpm_wave_actions_router)
     app.include_router(dpm_wave_lifecycle_actions_router)
