+++ src/app/routers/dpm_wave_lookup.py
@@ -1,20 +1,13 @@
-from fastapi import APIRouter, Path, Query
+from fastapi import APIRouter, Query
 
-from app.contracts.dpm_waves import DpmWaveErrorDetail, DpmWaveGatewayResponse
+from app.contracts.dpm_waves import DpmWaveGatewayResponse
 from app.middleware.correlation import correlation_id_var
-from app.routers.dpm_openapi import manage_upstream_error_responses
+from app.routers.dpm_wave_lookup_common import UPSTREAM_WAVE_LOOKUP_ERROR_RESPONSES
 from app.services.dpm_service_provider import dpm_wave_service
 
 router = APIRouter(
     prefix="/api/v1/dpm/command-center/waves",
     tags=["DPM Command Center"],
-)
-UPSTREAM_WAVE_LOOKUP_ERROR_RESPONSES = manage_upstream_error_responses(
-    error_model=DpmWaveErrorDetail,
-    not_found_description="lotus-manage could not find the requested rebalance wave.",
-    conflict_description="lotus-manage rejected the rebalance-wave request as conflicting.",
-    invalid_payload_description="lotus-manage rejected the rebalance-wave payload as invalid.",
-    unavailable_description="lotus-manage rebalance-wave authority is unavailable or degraded.",
 )
 
 
@@ -59,26 +52,5 @@ async def list_waves(
             "limit": limit,
             "offset": offset,
         },
-        correlation_id=correlation_id_var.get(),
-    )
-
-
-@router.get(
-    "/{wave_id}",
-    response_model=DpmWaveGatewayResponse,
-    summary="Get DPM rebalance wave",
-    description=(
-        "What: returns one durable manage-owned RFC-0041 rebalance wave. When: call this for "
-        "Workbench wave detail, PM review, CIO review, or operations drill-down. How: Gateway "
-        "preserves manage wave detail, item states, events, aggregate metrics, source refs, "
-        "supportability, proof-pack posture, and handoff posture without recomputation."
-    ),
-    responses=UPSTREAM_WAVE_LOOKUP_ERROR_RESPONSES,
-)
-async def get_wave(
-    wave_id: str = Path(..., description="Manage-owned rebalance-wave identifier."),
-) -> DpmWaveGatewayResponse:
-    return await dpm_wave_service().get_wave(
-        wave_id=wave_id,
         correlation_id=correlation_id_var.get(),
     )
